feat: 도감에 포획한 스톡몬이 없는 경우 ui 구현
