+++ src/app/books/page.tsx
@@ -1,39 +1,35 @@
 "use client";
 import CountHeader from "@/components/ui/books/CountHeader";
+import StockmonList from "@/components/ui/books/StockmonList";
 import CommonLayout from "@/components/ui/CommonLayout";
 import SearchBar from "@/components/ui/SearchBar";
 import { Skeleton } from "@/components/ui/Skeleton";
 import React, { useState } from "react";
 import useSWR from "swr";
 
-type StockmonItem = {
-  id: number;
-  name: string;
-  imgUrl: string;
-  count: number;
-  stockCode: number;
-  stockAveragePrice: number;
-};
+// const dummy = {
+//   stockmons: [
+//     {
+//       id: 1,
+//       name: "신한지주몬",
+//       imgUrl: "http~",
+//       count: 10,
+//       stockCode: 124125,
+//       stockAveragePrice: 145400,
+//     },
+//     {
+//       id: 2,
+//       name: "하나금융지주몬",
+//       imgUrl: "http~",
+//       count: 12,
+//       stockCode: 141125,
+//       stockAveragePrice: 235400,
+//     },
+//   ],
+// };
 
-const dummy: { stockmons: StockmonItem[] } = {
-  stockmons: [
-    {
-      id: 1,
-      name: "신한지주몬",
-      imgUrl: "http~",
-      count: 10,
-      stockCode: 124125,
-      stockAveragePrice: 145400,
-    },
-    {
-      id: 2,
-      name: "하나금융지주몬",
-      imgUrl: "http~",
-      count: 12,
-      stockCode: 141125,
-      stockAveragePrice: 235400,
-    },
-  ],
+const dummy = {
+  stockmons: [],
 };
 
 const fetcher = (url: string) => {
@@ -43,15 +39,14 @@ const fetcher = (url: string) => {
 
 export default function Books() {
   const { data, error } = useSWR("/api/user", fetcher);
-  let countHeader: React.ReactNode;
+  let countHeader: React.ReactNode = <CountHeader isLoading />;
+  let stockmonList: React.ReactNode = <StockmonList isLoading />;
 
   if (error) return <div>failed to load</div>;
-  if (!data) {
-    countHeader = <CountHeader isLoading />;
-  } else {
+  if (data && data.stockmons) {
     countHeader = <CountHeader count={data.stockmons.length} />;
+    stockmonList = <StockmonList stockmons={data.stockmons} />;
   }
-
   return (
     <CommonLayout
       title={"도감"}
@@ -63,14 +58,7 @@ export default function Books() {
         </div>
       }
     >
-      <div></div>
-      {/* <ul>
-        {data.stockmons.map((stockmon) => (
-          <li key={stockmon.id}>
-            <a href={`/books/${stockmon.id}`}>{stockmon.name}</a>
-          </li>
-        ))}
-      </ul> */}
+      {stockmonList}
     </CommonLayout>
   );
 }
