+++ src/components/ui/books/StockmonItem.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function StockmonItem() {
+  return <div>StockmonItem</div>;
+}
